Fix to use the date custom class
The myfirstgem.rb library file updates the call to the ShowTheDate class
by assigning a variable which is interpreted in the text when a request
is made to Rack. The RackUp config file will then send the response to
the web server (in this case Puma) and text with the current date will
display in your browser.

This site is currently hosted by a local computer and therefore the
request from the browser will be localhost:9292, or whichever port is
configured for your web server to send responses.